+++ build.go
@@ -27,3 +27,4 @@ import "runtime"
 const AppVersion = "0.0.0"
 const BuildOS = runtime.GOOS
 const BuildArch = runtime.GOARCH
+const BuildCompiler = "0.0.0"